+++ spag.py
@@ -13,84 +13,7 @@ import spag_files
 import spag_template
 from spag_remembers import SpagRemembers
 from common import ToughNoodles, update
-
-def determine_endpoint(f):
-    @functools.wraps(f)
-    def wrapper(*args, **kwargs):
-        if kwargs['endpoint'] is None:
-            try:
-                endpoint = spag_files.SpagEnvironment.get_env()['envvars']['endpoint']
-                kwargs['endpoint'] = endpoint
-            except ToughNoodles as e:
-                click.echo(str(e), err=True)
-                sys.exit(1)
-            except KeyError:
-                click.echo('Endpoint not set\n', err=True)
-                sys.exit(1)
-        return f(*args, **kwargs)
-    return wrapper
-
-def prepare_headers(f):
-    @functools.wraps(f)
-    def wrapper(*args, **kwargs):
-        header = kwargs['header']
-        if header is None or header == ():
-            try:
-                headers = spag_files.SpagEnvironment.get_env()['headers']
-                kwargs['header'] = headers
-            except ToughNoodles:
-                kwargs['header'] = None
-            except KeyError:
-                kwargs['header'] = None
-        else:
-            try:
-                # Headers come in as a tuple ('Header:Content', 'Header:Content')
-                env = spag_files.SpagEnvironment.get_env()
-                supplied_headers = {key: value.strip() for (key, value) in [h.split(':') for h in header]}
-                if 'headers' in env:
-                    headers = update(env['headers'], supplied_headers)
-                else:
-                    headers = supplied_headers
-                kwargs['header'] = headers
-            except ValueError:
-                click.echo("Error: Invalid header!", err=True)
-                sys.exit(1)
-            except KeyError:
-                kwargs['header'] = {key: value.strip() for (key, value) in [h.split(':') for h in header]}
-
-        return f(*args, **kwargs)
-    return wrapper
-
-def request_dir(f):
-    @functools.wraps(f)
-    def wrapper(*args, **kwargs):
-        if kwargs['dir'] is None:
-            try:
-                d = spag_files.SpagEnvironment.get_env()['envvars']['dir']
-                kwargs['dir'] = d
-            except ToughNoodles as e:
-                click.echo(str(e), err=True)
-                sys.exit(1)
-            except KeyError:
-                click.echo('Request directory not set\n', err=True)
-                sys.exit(1)
-        return f(*args, **kwargs)
-    return wrapper
-
-def common_request_args(f):
-    @click.option('--endpoint', '-e', metavar='<endpoint>',
-                  default=None, help='Manually override the endpoint')
-    @click.option('--header', '-H', metavar = '<header>', multiple=True,
-                  default=None, help='Header in the form Key:Value')
-    @click.option('--show-headers', '-h', is_flag=True,
-                  help='Prints the headers along with the response body')
-    @click.option('--data', '-d', required=False, help='the request data')
-    @determine_endpoint
-    @prepare_headers
-    @functools.wraps(f)
-    def wrapper(*args, **kwargs):
-        return f(*args, **kwargs)
-    return wrapper
+import decorators as dec
 
 
 @click.group()
@@ -109,7 +32,7 @@ def show_response(resp, show_headers):
 
 @cli.command('get')
 @click.argument('resource')
-@common_request_args
+@dec.common_request_args
 def get(resource, endpoint=None, data=None, header=None, show_headers=False):
     """HTTP GET"""
     uri = endpoint + resource
@@ -120,7 +43,7 @@ def get(resource, endpoint=None, data=None, header=None, show_headers=False):
 
 @cli.command('post')
 @click.argument('resource')
-@common_request_args
+@dec.common_request_args
 def post(resource, endpoint=None, data=None, header=None, show_headers=False):
     """HTTP POST"""
     uri = endpoint + resource
@@ -131,7 +54,7 @@ def post(resource, endpoint=None, data=None, header=None, show_headers=False):
 
 @cli.command('put')
 @click.argument('resource')
-@common_request_args
+@dec.common_request_args
 def put(resource, endpoint=None, data=None, header=None, show_headers=False):
     """HTTP PUT"""
     uri = endpoint + resource
@@ -142,7 +65,7 @@ def put(resource, endpoint=None, data=None, header=None, show_headers=False):
 
 @cli.command('patch')
 @click.argument('resource')
-@common_request_args
+@dec.common_request_args
 def patch(resource, endpoint=None, data=None, header=None, show_headers=False):
     """HTTP PATCH"""
     uri = endpoint + resource
@@ -153,7 +76,7 @@ def patch(resource, endpoint=None, data=None, header=None, show_headers=False):
 
 @cli.command('delete')
 @click.argument('resource')
-@common_request_args
+@dec.common_request_args
 def delete(resource, endpoint=None, data=None, header=None, show_headers=False):
     """HTTP DELETE"""
     uri = endpoint + resource
@@ -165,8 +88,8 @@ def delete(resource, endpoint=None, data=None, header=None, show_headers=False):
 
 @cli.command('request')
 @click.argument('name', required=False)
-@common_request_args
-@request_dir
+@dec.common_request_args
+@dec.request_dir
 @click.option('--dir', required=False,
               help='the dir to search for request files')
 @click.option('--show', required=False, is_flag=True,
